added model file
